Add /bump-now-playing, improve advance skipping
