+++ lyra/src/lavalink/mod.rs
@@ -8,10 +8,9 @@ pub use self::{
     model::{
         ClientAndGuildIdAware, ClientAware, ClientData, Connection, CorrectPlaylistInfo,
         CorrectTrackInfo, DelegateMethods, Event, EventRecvResult, IndexerType, Lavalink,
-        OwnedPlayerData, Pitch, PlayerDataRead, PlayerDataWrite, Queue, QueueItem, RepeatMode,
-        UnwrappedData, UnwrappedPlayerInfoUri, wait_for_with,
+        NowPlayingData, OwnedPlayerData, Pitch, PlayerDataRead, PlayerDataWrite, Queue, QueueItem,
+        RepeatMode, UnwrappedData, UnwrappedPlayerInfoUri, wait_for_with,
     },
     plugin::lava_src::PluginInfo,
     process::handlers,
-    track::delete_now_playing_message,
 };
