+++ lyra/src/lavalink/model/connection.rs
@@ -1,18 +1,27 @@
-use lyra_ext::nested_transpose::NestedTranspose;
-use tokio::sync::{broadcast, watch};
-use twilight_model::id::{
-    Id,
-    marker::{ChannelMarker, UserMarker},
+use std::{
+    collections::HashMap,
+    pin::Pin,
+    task::{Context, Poll},
 };
 
-use crate::{command::poll::Poll, core::r#const};
+use futures::FutureExt;
+use lyra_ext::nested_transpose::NestedTranspose;
+use tokio::sync::{broadcast, mpsc, oneshot, watch};
+use twilight_model::id::{
+    Id,
+    marker::{ChannelMarker, GuildMarker, UserMarker},
+};
+
+use crate::{command::poll::Poll as PlayerPoll, core::r#const, error::UnrecognisedConnection};
+
+use super::Lavalink;
 
 #[derive(Debug)]
 pub struct Connection {
     pub channel_id: Id<ChannelMarker>,
     pub text_channel_id: Id<ChannelMarker>,
     pub mute: bool,
-    poll: Option<Poll>,
+    poll: Option<PlayerPoll>,
     change: watch::Sender<()>,
     event_sender: broadcast::Sender<Event>,
 }
@@ -36,11 +45,11 @@ impl Connection {
         self.change.subscribe()
     }
 
-    pub const fn poll(&self) -> Option<&Poll> {
+    pub const fn poll(&self) -> Option<&PlayerPoll> {
         self.poll.as_ref()
     }
 
-    pub const fn set_poll(&mut self, poll: Option<Poll>) {
+    pub const fn set_poll(&mut self, poll: Option<PlayerPoll>) {
         self.poll = poll;
     }
 
@@ -51,7 +60,7 @@ impl Connection {
 
     /// Notify the connection to trigger a change.
     pub fn notify_change(&self) {
-        tracing::trace!("notified connection change");
+        tracing::debug!("notified connection change");
         self.change.send(()).ok();
     }
 
@@ -145,4 +154,254 @@ pub async fn wait_for_with(
     });
 
     Ok(event.await.transpose()?.ok())
+}
+
+type Response<T> = oneshot::Sender<Result<T, UnrecognisedConnection>>;
+
+pub(super) enum Instruction {
+    /// Insert a new connection
+    Insert(Id<GuildMarker>, Connection),
+    /// Remove a connection
+    Remove(Id<GuildMarker>),
+    /// Query if a connection exists
+    Exists(Id<GuildMarker>, oneshot::Sender<bool>),
+    /// Dispatch an event to a connection
+    Dispatch(Id<GuildMarker>, Event, Response<()>),
+    /// Subscribe to events from a connection
+    Subscribe(Id<GuildMarker>, Response<broadcast::Receiver<Event>>),
+    /// Notify a connection of a change
+    NotifyChange(Id<GuildMarker>, Response<()>),
+    /// Subscribe to changes from a connection
+    SubscribeOnChange(Id<GuildMarker>, Response<watch::Receiver<()>>),
+    /// Toggle mute
+    ///
+    /// Returns the current mute state if successful
+    ToggleMute(Id<GuildMarker>, Response<bool>),
+    /// Set mute
+    SetMute(Id<GuildMarker>, bool, Response<()>),
+    /// Set the channel for a connection
+    SetChannel(Id<GuildMarker>, Id<ChannelMarker>, Response<()>),
+    /// Set the text channel for a connection
+    SetTextChannel(Id<GuildMarker>, Id<ChannelMarker>, Response<()>),
+    /// Get basic connection info
+    Head(Id<GuildMarker>, Response<ConnectionHead>),
+    /// Get the connection poll info
+    GetPoll(Id<GuildMarker>, Response<Option<PlayerPoll>>),
+    /// Set the connection poll info
+    SetPoll(Id<GuildMarker>, Option<PlayerPoll>, Response<()>),
+}
+
+/// The result of a future that waits for a value to be sent
+pub struct Awaitable<T> {
+    receiver: oneshot::Receiver<T>,
+}
+
+impl<T> Future for Awaitable<T> {
+    type Output = T;
+
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        match self.receiver.poll_unpin(cx) {
+            Poll::Ready(Ok(result)) => Poll::Ready(result),
+            Poll::Ready(Err(_)) => panic!("Actor sent no result (This is a bug)"),
+            Poll::Pending => Poll::Pending,
+        }
+    }
+}
+
+pub(super) struct ConnectionsActor {
+    connections: HashMap<Id<GuildMarker>, Connection>,
+    receiver: mpsc::UnboundedReceiver<Instruction>,
+}
+
+impl ConnectionsActor {
+    pub fn new(receiver: mpsc::UnboundedReceiver<Instruction>) -> Self {
+        Self {
+            connections: HashMap::new(),
+            receiver,
+        }
+    }
+
+    fn with_connection<T>(
+        &self,
+        guild_id: Id<GuildMarker>,
+        sender: oneshot::Sender<Result<T, UnrecognisedConnection>>,
+        f: impl FnOnce(&Connection) -> T,
+    ) {
+        if let Some(connection) = self.connections.get(&guild_id) {
+            let result = f(connection);
+            let _ = sender.send(Ok(result));
+        } else {
+            let _ = sender.send(Err(UnrecognisedConnection));
+        }
+    }
+
+    fn with_connection_mut<T>(
+        &mut self,
+        guild_id: Id<GuildMarker>,
+        sender: oneshot::Sender<Result<T, UnrecognisedConnection>>,
+        f: impl FnOnce(&mut Connection) -> T,
+    ) {
+        if let Some(connection) = self.connections.get_mut(&guild_id) {
+            let result = f(connection);
+            let _ = sender.send(Ok(result));
+        } else {
+            let _ = sender.send(Err(UnrecognisedConnection));
+        }
+    }
+
+    pub async fn run(&mut self) {
+        while let Some(instruction) = self.receiver.recv().await {
+            match instruction {
+                Instruction::Insert(guild_id, connection) => {
+                    self.connections.insert(guild_id, connection);
+                }
+                Instruction::Remove(guild_id) => {
+                    self.connections.remove(&guild_id);
+                }
+                Instruction::Exists(guild_id, sender) => {
+                    let exists = self.connections.contains_key(&guild_id);
+                    let _ = sender.send(exists);
+                }
+                Instruction::Dispatch(guild_id, event, sender) => {
+                    self.with_connection(guild_id, sender, |c| c.dispatch(event));
+                }
+                Instruction::Subscribe(guild_id, sender) => {
+                    self.with_connection(guild_id, sender, Connection::subscribe);
+                }
+                Instruction::NotifyChange(guild_id, sender) => {
+                    self.with_connection(guild_id, sender, Connection::notify_change);
+                }
+                Instruction::SubscribeOnChange(guild_id, sender) => {
+                    self.with_connection(guild_id, sender, Connection::subscribe_on_changed);
+                }
+                Instruction::ToggleMute(guild_id, sender) => {
+                    self.with_connection_mut(guild_id, sender, |c| {
+                        c.mute = !c.mute;
+                        c.mute
+                    });
+                }
+                Instruction::SetMute(guild_id, mute, sender) => {
+                    self.with_connection_mut(guild_id, sender, |c| {
+                        c.mute = mute;
+                    });
+                }
+                Instruction::SetChannel(guild_id, channel_id, sender) => {
+                    self.with_connection_mut(guild_id, sender, |c| {
+                        c.channel_id = channel_id;
+                    });
+                }
+                Instruction::SetTextChannel(guild_id, channel_id, sender) => {
+                    self.with_connection_mut(guild_id, sender, |c| {
+                        c.text_channel_id = channel_id;
+                    });
+                }
+                Instruction::Head(guild_id, sender) => {
+                    self.with_connection(guild_id, sender, |c| c.into());
+                }
+                Instruction::GetPoll(guild_id, sender) => {
+                    self.with_connection(guild_id, sender, |c| c.poll().cloned());
+                }
+                Instruction::SetPoll(id, poll, sender) => {
+                    self.with_connection_mut(id, sender, |c| {
+                        c.set_poll(poll);
+                    });
+                }
+            }
+        }
+    }
+}
+
+/// Represents a connection to the lavalink server
+pub struct ConnectionHandle<'a> {
+    pub(super) parent: &'a Lavalink,
+    pub(super) guild_id: Id<GuildMarker>,
+}
+
+impl ConnectionHandle<'_> {
+    fn send_instruction(&self, instruction: Instruction) {
+        self.parent
+            .sender
+            .as_ref()
+            .expect("Lavalink was not started")
+            .send(instruction)
+            .expect("Lavalink instruction sender must not be closed");
+    }
+
+    fn call_awaitable<T>(
+        &self,
+        f: impl FnOnce(Response<T>) -> Instruction,
+    ) -> Awaitable<Result<T, UnrecognisedConnection>> {
+        let (sender, receiver) = oneshot::channel();
+        self.send_instruction(f(sender));
+        Awaitable { receiver }
+    }
+
+    async fn call_result<T>(
+        &self,
+        f: impl FnOnce(Response<T>) -> Instruction,
+    ) -> Result<T, UnrecognisedConnection> {
+        let (sender, receiver) = oneshot::channel();
+        self.send_instruction(f(sender));
+        receiver
+            .await
+            .expect("Lavalink connection sender must not be closed")
+    }
+
+    pub fn dispatch(&self, event: Event) -> Awaitable<Result<(), UnrecognisedConnection>> {
+        self.call_awaitable(|sender| Instruction::Dispatch(self.guild_id, event, sender))
+    }
+
+    pub fn notify_change(&self) -> Awaitable<Result<(), UnrecognisedConnection>> {
+        self.call_awaitable(|sender| Instruction::NotifyChange(self.guild_id, sender))
+    }
+
+    pub fn toggle_mute(&self) -> Awaitable<Result<bool, UnrecognisedConnection>> {
+        self.call_awaitable(|sender| Instruction::ToggleMute(self.guild_id, sender))
+    }
+
+    pub fn set_mute(&self, mute: bool) -> Awaitable<Result<(), UnrecognisedConnection>> {
+        self.call_awaitable(|sender| Instruction::SetMute(self.guild_id, mute, sender))
+    }
+
+    pub fn set_channel(
+        &self,
+        channel_id: Id<ChannelMarker>,
+    ) -> Awaitable<Result<(), UnrecognisedConnection>> {
+        self.call_awaitable(|sender| Instruction::SetChannel(self.guild_id, channel_id, sender))
+    }
+
+    pub fn set_text_channel(
+        &self,
+        channel_id: Id<ChannelMarker>,
+    ) -> Awaitable<Result<(), UnrecognisedConnection>> {
+        self.call_awaitable(|sender| Instruction::SetTextChannel(self.guild_id, channel_id, sender))
+    }
+
+    pub fn set_poll(&self, poll: PlayerPoll) -> Awaitable<Result<(), UnrecognisedConnection>> {
+        self.call_awaitable(|sender| Instruction::SetPoll(self.guild_id, Some(poll), sender))
+    }
+
+    pub fn reset_poll(&self) -> Awaitable<Result<(), UnrecognisedConnection>> {
+        self.call_awaitable(|sender| Instruction::SetPoll(self.guild_id, None, sender))
+    }
+
+    pub async fn get_poll(&self) -> Result<Option<PlayerPoll>, UnrecognisedConnection> {
+        self.call_result(|sender| Instruction::GetPoll(self.guild_id, sender))
+            .await
+    }
+
+    pub async fn get_head(&self) -> Result<ConnectionHead, UnrecognisedConnection> {
+        self.call_result(|sender| Instruction::Head(self.guild_id, sender))
+            .await
+    }
+
+    pub async fn subscribe(&self) -> Result<broadcast::Receiver<Event>, UnrecognisedConnection> {
+        self.call_result(|sender| Instruction::Subscribe(self.guild_id, sender))
+            .await
+    }
+
+    pub async fn subscribe_on_change(&self) -> Result<watch::Receiver<()>, UnrecognisedConnection> {
+        self.call_result(|sender| Instruction::SubscribeOnChange(self.guild_id, sender))
+            .await
+    }
 }
