+++ lyra/src/lavalink/track/end.rs
@@ -1,9 +1,8 @@
 use lavalink_rs::{client::LavalinkClient, model::events::TrackEnd};
 
 use crate::{
-    core::model::HttpAware,
     error::lavalink::ProcessResult,
-    lavalink::{CorrectTrackInfo, UnwrappedData, model::PlayerData},
+    lavalink::{CorrectTrackInfo, UnwrappedData},
 };
 
 #[tracing::instrument(err, skip_all, name = "track_end")]
@@ -20,17 +19,22 @@ pub(super) async fn impl_end(
     );
 
     let Some(player) = lavalink.get_player_context(guild_id) else {
-        tracing::trace!(?guild_id, "track ended via forced disconnection");
+        tracing::debug!(?guild_id, "track ended via forced disconnection");
 
         return Ok(());
     };
     let data = player.data_unwrapped();
 
-    delete_now_playing_message(lavalink.data_unwrapped().as_ref(), &data).await;
+    data.write()
+        .await
+        .delete_now_playing_message(lavalink.data_unwrapped().as_ref())
+        .await;
 
     let data_r = data.read().await;
-    if data_r.queue().not_advance_locked().await {
-        tracing::trace!(?guild_id, "track ended normally");
+    if data_r.queue().skip_advance().await {
+        tracing::debug!(?guild_id, "track ended forcefully");
+    } else {
+        tracing::debug!(?guild_id, "track ended normally");
 
         drop(data_r);
         let mut data_w = data.write().await;
@@ -41,18 +45,7 @@ pub(super) async fn impl_end(
             player.play_now(item.data()).await?;
         }
         drop(data_w);
-    } else {
-        tracing::trace!(?guild_id, "track ended forcefully");
     }
 
     Ok(())
-}
-
-pub async fn delete_now_playing_message(cx: &(impl HttpAware + Sync), data: &PlayerData) {
-    let mut data_w = data.write().await;
-    if let Some(message) = data_w.take_now_playing_message() {
-        let channel_id = message.channel_id();
-        let _ = cx.http().delete_message(channel_id, message.id()).await;
-    }
-    drop(data_w);
 }
